feat(filter-products): add use filters composable
